[SUBMITTED 20171019] x86: don't include asm/x86_init.h in asm/setup.h
This is a preparation to allow using 'struct timespec64' in the
asm/x86_init.h.  Unfortunately, we can't use a forward declaration
for timespec64 since it is defined as a macro on 64-bit architectures,
and including linux/time64.h breaks compilation of arch/x86/boot/,
which runs in realmode and can't use many of the regular kernel
headers.

As a workaround, I stop including asm/x86_init.h. This works fine
for the realmode boot code since it does not require any of the
x86_init.h contents, and setup.h doesn't either. However, a couple
of other files that do need x86_init.h used to rely on it being
included indirectly, so I have to put an explicit include in there
now.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>