staging: lustre: avoid timeval on uapi (open-coded)
lustre walks a linked list of lstcon_rpc_ent structures in user
space. These structures contain a 'struct timeval', which is defined by
glibc and will break the structure layout when built using a 32-bit libc
with 64-bit time_t.

This changes the structure to use the newly introduced
__kernel_old_timeval type that makes it a fixed layout independent
of the C library version.

Since jiffies_to_timeval() no longer works on that structure, and
is deprecated, I'm changing the code to use the jiffies_to_timespec64()
function instead, and manyally convert into the user space type, with
a comment to explain that there is no y2038 problem in lustre itself.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>