[SUBMITTED 20180118] PCI: dwc: dra7xx: add back CONFIG_PCI dependency for endpoint
It was a nice idea to split out the PCI host and endpoint mode configuration
into separate options. Unfortunately it doesn't build:

drivers/pci/dwc/pci-dra7xx.c:229:11: error: 'pci_irqd_intx_xlate' undeclared here (not in a function)

This is certainly a fixable problem, but since it's clear that this
configuration was never tested, let's maybe revert back to the
dependency for now, until it can be done in a way that works
better.

Fixes: b052835c6385 ("PCI: dwc: dra7xx: Refactor Kconfig and Makefile handling for host/ep mode")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>