powerpc: use time64_t in read_persistent_clock
Looking through the remaining users of the deprecated mktime()
function, I found the powerpc rtc handlers, which use it in
place of rtc_tm_to_time64(). Unfortunately rtc_tm_to_time64() and
rtc_time64_to_tm() are only available when CONFIG_RTC_CLASS is enabled,
but that is optional on powerpc unlike most architectures.

To clean this up, I'm changing over the read_persistent_clock()
function to the read_persistent_clock64() variant, and change
all the platform specific handlers along with it.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>