parisc: time: convert to read_persistent_clock64
The read_persistent_clock() interface is deprecated in favor of
read_persistent_clock64(), which supports times beyond y2038.

We could in theory remove the read_persistent_clock() call entirely
here and just rely on the rtc_generic driver to read the time, but
that has the disadvantage of losing the microsecond portion of the
time returned by the firmware.

Simply using read_persistent_clock64() extends the available range
for the RTC to year 2106 on 32-bit kernels and beyond that on 64-bit.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>