[SUBMITTED 20180220] drm/msm/hdmi: avoid -Wrestrict warning
gcc-8 with link time optimizations warns that we might be
copying a fixed length object into a shorted fixed-length
buffer:

In function 'hdmi_spd_infoframe_pack.constprop',
    inlined from 'hdmi_infoframe_pack.constprop' at drivers/video/hdmi.c:434:12,
    inlined from 'msm_hdmi_config_avi_infoframe' at drivers/gpu/drm/msm/hdmi/hdmi_bridge.c:106:8,
    inlined from 'msm_hdmi_bridge_pre_enable' at drivers/gpu/drm/msm/hdmi/hdmi_bridge.c:165:4:
drivers/video/hdmi.c:214:2: error: 'memcpy' writing 16 bytes into a region of size 5 overflows the destination [-Werror=stringop-overflow=]
  memcpy(ptr + 8, frame->product, sizeof(frame->product));

In this case, it seems that gcc is wrong, but it's hard to blame it,
since the code path is rather non-obvious: we assign the frame type in
drm_hdmi_avi_infoframe_from_display_mode() and then have a switch/case
statement in hdmi_avi_infoframe_pack() that checks for the type to decide
which other function to call that then copies the correct buffer.

It is very easy to work around it by calling the more specific
hdmi_avi_infoframe_pack() helper instead, which always takes the
correct size.

Adding Martin Sebor for reference, as he worked on similar false-positive
warnings that unlike this one should be addressed by improving gcc.

Cc: Martin Sebor <msebor@gcc.gnu.org>
Cc: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=84095
Signed-off-by: Arnd Bergmann <arnd@arndb.de>