[SUBMITTED 20180313] platform/x86: dell-smbios: fix DCDBAS dependency, better
The rework of the dell laptop modules caused a link error:

drivers/platform/x86/dell-smbios-smm.o: In function `dell_smbios_smm_call':
dell-smbios-smm.c:(.text+0x77): undefined reference to `dcdbas_smi_request'

We already have a fix for that, but this was only a partial solution,
since CONFIG_DELL_LAPTOP and CONFIG_DELL_WMI select CONFIG_DELL_SMBIOS
regardless of that dependency.

We can either add the same 'DCDBAS || DCDBAS=n' dependency for those two
as well, or replace the 'select' with a 'depends on'. The second approach
seems cleaner, so I'm picking that here.

Fixes: 32d7b19bad96 ("platform/x86: dell-smbios: Resolve dependency error on DCDBAS")
Fixes: 25d47027e100 ("platform/x86: dell-smbios: Link all dell-smbios-* modules together")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>