ALSA: cmipci: avoid string overflow warning
Recent gcc-8 snapshots thinks the model string could overflow because
it has information about the range of the input integer value:

sound/pci/cmipci.c:3157:21: note: directive argument in the range [-2147483648, 67]
   sprintf(modelstr, " (model %d)", cm->chip_version);
                     ^~~~~~~~~~~~~
sound/pci/cmipci.c:3157:3: note: 'sprintf' output between 11 and 21 bytes into a destination of size 16
   sprintf(modelstr, " (model %d)", cm->chip_version);

This is a false positive, as we can tell that the version is always set
to a constant number between 0 and 68. It's unclear to me whether we
really want that warning here, or if gcc should be a bit smarter about
the range check when it encounters a signed integer with an upper bound
but no lower bound. Without knowing the upper bound, gcc won't warn
here, apparently to avoid excessive false positives.

Cc: Martin Sebor <msebor@gcc.gnu.org>
Signed-off-by: Arnd Bergmann <arnd@arndb.de>