x86: olpc: fix inconsistent MFD_CS5535 configuration
The GPIO_CS5535 driver depends on the MFD_CS5535 driver, but the former
is selected in places where the latter is not:

WARNING: unmet direct dependencies detected for GPIO_CS5535
  Depends on [m]: GPIOLIB [=y] && (X86 [=y] || MIPS || COMPILE_TEST [=y]) && MFD_CS5535 [=m]
  Selected by [y]:
  - OLPC_XO1_SCI [=y] && X86_32 [=y] && OLPC [=y] && OLPC_XO1_PM [=y] && INPUT [=y]=y

The 'select MFD_CORE' part is completely redundant, since we already
depend on MFD_CS5535 here. Instead of selecting GPIO_CS5535, we can use
'depends on GPIO_CS5535'. Ideally, the private symbols exported by that
cs5535 gpio driver would just be converted to gpiolib interfaces so we
could expletely avoid this dependency.

Signed-off-by: Arnd Bergmann <arnd@arndb.de>