[SUBMITTED 20180726] sh: remove board_time_init() callback
The only remaining user of board_time_init() is the of-generic
machine, and that just calls the global timer_init() function.
Calling that one has no effect on non-DT platforms, so we can
simply call it unconditionally in place of board_time_init().

Signed-off-by: Arnd Bergmann <arnd@arndb.de>
---
v2: fix a typo as pointed out by the kbuild bot