[SUBMITTED 20180525] drm/bridge: fix dependency for lvds-encoder
The DRM panel bridge code is built into the kms helpers module, so we
get a link error when trying to use it from a built-in driver while the
kms helper is a loadable module:

drivers/gpu/drm/bridge/lvds-encoder.o: In function `lvds_encoder_probe':
lvds-encoder.c:(.text+0x124): undefined reference to `devm_drm_panel_bridge_add'

This adds a the same dependency in the lvds-encoder that we use for all
the other users of the panel bridge. I did not bisect the problem, but
from inspection it seems to date back to the patch that separated out
the panel bridge from lvds encoder.

Fixes: 13dfc0540a57 ("drm/bridge: Refactor out the panel wrapper from the lvds-encoder bridge.")
Signed-off-by: Arnd Bergmann <arnd@arndb.de>